Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
urllib3==2.1.0
pandas>=2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
pydantic==2.5.0
alembic==1.14.0
httpx==0.27.2
pytest==8.3.4
